Instance and term separation
